Lee Elliot:
I've attached an archive of the experimental SU-37 that I've been
working on.

I've set the status to ' experimental' as I think that's the best
description  - it's all very much experimental to me :)

I've tried a number of different ways of organising the FCS stuff
and although I think I've settled upon a final policy/scheme I'm
only half way through making everything compliant with it - in
other words, bit's of it are a mess :)

The thrust vectoring works but I've had to use a bit of a
work-around in YASim to do it - I had to rotate the nozzles 720
deg so that I could get a contiguous +/- 15 deg range, centered
at 360 deg to do it :)  The canards appear to have no
aerodynamic effect at all and looking at the source I'm not sure
if variable incidence has actually been implemented even though
there's a control axis for it - I'll have to bug Andy R about
these at some point.

It's been interesting stuff to work on and has highlighted a
number of desirable features e.g. a method of routing control
input outside of using Nasal (because of timing/rate factors)
and switchable filters whose parameters can be changed.

For example, I'd like to add a scaled pilot input function for
the FCS 'direct' mode so I can vary the sensitivity of the pilot
input and being able to vary the gain on a filter would be very
handy for this - simply scaling the raw input in Nasal works
over the range of the scaling factor but also limits the range.

The 3d model still needs some work to tidy up and refine a few
bits, and it still needs texturing, of course, but I'll probably
be spending more time working on the control stuff... because
it's more interesting:)