Lee Elliott:
here's an update for the SU-37 cvs version.  It includes the stuff from an
earlier update that hasn't made it into cvs yet so if that turns up it's safe
to ignore it.

This just about completes the Nasal re-writes and re-organisation.

I've also added a cubed-input function (when in direct stick mode) and a hack
that will often result in a spin:)

There are a couple of files to remove as well

SU-37/Nasal/SU-37-FCS-slave-listeners.nas
SU-37/Nasal/SU-37-FCS-velocities-listeners.nas